Changed to not reprocess files that have already been processed.

+++ organize_rendering_data.py
@@ -25,13 +25,14 @@ for scene in l:
                 frame = os.path.splitext(os.path.basename(exr_path))[0][1:]    # '**/**/Z0104.exr' -> '0104'
                 os.makedirs(new_dir, exist_ok=True)
 
-                # EXR(depth) -> npy(disparity)
-                img, Z = readEXR(exr_path)
-                for scene_id, num in lt:
-                    if scene_id in old_dir:
-                        focal_length = float(num)
-                disp = depth2disp(Z, baseline=0.01, focal_length=focal_length)
-                np.save(f'{new_dir}/{frame}', disp)
+                if (not os.path.exists(f'{new_dir}/{frame}.npy')) and (not os.path.exists(f'{new_dir}/{frame}.png')):
+                    # EXR(depth) -> npy(disparity)
+                    img, Z = readEXR(exr_path)
+                    for scene_id, num in lt:
+                        if scene_id in old_dir:
+                            focal_length = float(num)
+                    disp = depth2disp(Z, baseline=0.01, focal_length=focal_length)
+                    np.save(f'{new_dir}/{frame}', disp)
 
-                # copy .png
-                shutil.copy(f'{old_dir}/{frame}.png', new_dir)
+                    # copy .png
+                    shutil.copy(f'{old_dir}/{frame}.png', new_dir)
